Move everything into the scala.colleciton.compat package
Fixes https://github.com/scala/scala-collection-compat/issues/30

Using `scala.collection.compat.imm` (not `.immutable`) to avoid
name clashes with `scala.collection.immutable`.
