+++ src/main/scala-2.13/scala/collection/compat/imm/package.scala
@@ -1,0 +1,6 @@
+package scala.collection.compat
+
+package object imm {
+  type ArraySeq[+T] = scala.collection.immutable.ArraySeq[T]
+  val ArraySeq      = scala.collection.immutable.ArraySeq
+}
